renderPassOptions : Fix `option:renderPass:type` presets
This was broken when we introduced the `target` argument to `Metadata::ValueFunction` in b059de67716759d5aed095981179057d854e8b7a.

Rather than update the `RenderPassTypeAdaptorUI.renderPassTypePresetNames()` and `renderPassTypePresetValues()` functions to accept an unused `target` argument, I've addressed the existing todo in `startup/GafferSceneUI/renderPassOptions.py` and consolidated it into `startup/GafferScene/renderPassOptions.py`. `renderPassTypePresetNames()` and `renderPassTypePresetValues()` now return the metadata values, though they now might be worth removing entirely?
